Changed to use either thrading or multiprocessing
The accumulator and pipeline were originiaaly run using threading.
The threading and multiprocessing APIs are pretty similar, so the control has now been made
generic - the accumulator and pipeline can now either be run as threads or processes,
specified as a user input to the run_cal.py script (default is multiprocessing).

One significant difference between the threading and multiprocessing implimentations
is that for multiprocessing we need to set up shared memory for the data buffers.

The hope is that using multiprocessing over threading will fix the bug where things fall
over if the load gets bigger than the machine can handle, and packets are missed in the
accumulator because the pipeline is hogging the proccesing (this is my current
diagnosis of the problem!)

But I decided to keep both threading and multiprocessing as options so we can
test and benchmark both with more realistic data later.