Changed from using track_start_indices to max_index
The buffer used to contain a 'track_start_indices' key which was a list
of the start indices of each scan. The list was accumulated in the accumulator
and used in reduction.py to determine which data in the buffer falls into which scans.

I changed this so that the buffer now contains a 'max_index' key which is just the
maximum index in the buffer that was filled in the accumulator.
reduction.py then checks the activity to separate ut scans.

This change was made because keeping track of which timestamps fall into slews or
tracks and is unnecessary complexity for the accumulator.
(It resulted in some slew timestamps being missed and not ending up in the L1 stream.
Which I initially thought was fine, but if we want to make files with all of the
L0->L1 timestamps, we need to make sure all of the timestamps end up in L1).