Remove unneeded squeezes
The k_fit method promises to return an array of shape
(num_pols, num_ants). If num_pols == 1, the first dimension is
currently squeezed away, leading to grief further down the line
when the K solutions are applied.

Remove all squeezes, as the coarse_k shape is guaranteed to be
(num_ants,) and kdelay is then a list of coarse_k's of length
num_pols. The np.atleast_2d() simply highlights this, but a straight
np.array(kdelay) could also have worked.