Support for weights_channel and optimise accumulator
I had to fix up the simulator to split weights into 'weights' and
'weights_channel' SPEAD items to match the L0 stream, and from there I
needed to support that split in the accumulator. For now weights are
reconstituted as they're received, rather than lazily.

The copies from the SPEAD items into the accumulation buffers is
optimised by eliminating a copy from the permutation process.