Keep tasks running across observations
This allows the receiver to continue receiving the next observation
while the previous observation is still in the pipeline.

There are still a few issues, some of them inherited:
- If an observation ends with an empty scan, the observation end time is
  set to time.time(), rather than the last timestamp of the previous
  scan.
- Graceful shutdown between observations writes a possibly empty
  calibration report. The accumulator probably needs to detect empty
  observations.

Ctrl-C has been changed to do graceful rather than hard shutdown, mostly
so that it can be tested. Eventually SIGTERM will do graceful shutdown
too, but for now that will break the katsdpcontroller model.