Don't skip scans with fewer than three dumps
Since activity parsing has improved with the introduction of the katdal
sensor framework, there should be no need for this check. If problems
recur we can always revisit this.

This will please commissioners who are forced to spend at least five
dumps (i.e. more than 40 seconds for 8-second dumps) on target, which
is unnecessarily long for phasing up on a strong source.