Convert double buffer to ring buffer
This makes more efficient use of the available memory, and allows a
more-or-less arbitrary number of batches to be queued up for the
pipeline, instead of at most 1. This should address SR-887.

This is achieved by push dask further up the stack, into the pipeline
task. The accumulator sends a list of slot indices, and the pipeline
task assembles the data into a dask array.

A side effect is that the --buffer-maxsize argument now represents the
whole buffer, rather than one of the two buffers, and should be adjusted
in katsdpcontroller.

As part of this change, some of the code that threw away the weights was
fixed up to actually use them. The L1 transmit code was updated but
hasn't been tested, so might be broken - this can be addressed when the
flag-only stream gets implemented.