Move bp_fit to calprocs_dask and dask-ify
The main reason is to parallelise execution of stefcal. It could still
be a bit of a memory hog: see https://github.com/dask/dask/issues/874

The phase normalisation is changed a bit to avoid using a median, which
is not available in dask. Instead, the angles are wrapped such that the
branch cut is on the side opposite an arbitrary data point, and then the
mean is taken. Provided all the phases are clustered within pi of each
other this should be just as good, but it is less robust to a bad
outlier (particularly since the "arbitrary" point chosen is the minimum
initial angle).