Gracefully handle exceptions from the pipeline
Previously an exception in run_pipeline would wedge cal, since it would
just stop processing anything. With the latest changes on
katsdpcontroller, it now also wedges deconfiguration of the entire
subarray, because the program block never leaves the pipeline.

An exception is now logged, and there is also a sensor to indicate the
number of exceptions, as a convenient way to check if this has happened
in the past without picking through the logs.