Improve elevation calculation and timestamp formatting
Previously the report reconstructed timestamps from the median time of the scan and the number of increments. The code has been altered to pass the timestamps to the report directly, circumventing the reconstruction process.

Separate helper functions to format the times in the report and in the plot axes have been added.