Make the simulator live again
This is almost a total rewrite, to fix the weird reverse-inheritence
pattern, the unnecessary inheritence from specific katdal dataset
classes, the conflation of katdal with HDF5 and other gripes.

The MS support is almost certainly broken; I've only made things
work for katdal files for now.

SimData is now a base class, from which SimDataKatdal and SimDataMS
derive to provide the implementation of data loading. There is no longer
an independent file type detector: we just try to create a SimDataKatdal
first, and if it fails to open the file we fall back to trying
SimDataMS.

The simulator no longer has anything to do with parameter files;
parameters are up to how you run run_cal.py. Subsetting of channels
the file is done by passing --bchan and --echan to both sim_ts and
sim_data (and be sure they match!) I haven't put much work into the
feature since cal only works if it gets exactly L band anyway.

Print statements have been replaced by using the logging module, so it
is possible to see the times at which events took place.

I've renamed 'ts' to 'telstate' in most places that it is a telstate, to
avoid confusion with the use of 'ts' to mean timestamp.