Fill in data_lost when a dump goes missing
This is to ensure (as well as possible) that all the cal servers make
decisions in lockstep (about where to flush things), even if one of them
loses some data. This depends on activity and target sensors being
updated timeously, to avoid the situation where one cal server sees
stale data while another sees fresh data.

Missing timestamps are synthesised from the int_time, but might not
perfectly match the value that would be computed by ingest (rounding).
That would require some changes to ingest e.g. to set the L0 sync time
to match the time of dump 0, or to expose ADC timestamps.

This still needs some unit tests to exercise the corner cases.