Factor out _ensure_slots function
This function handles assigning new dump indices to slots, in
preparation for SR-1098. This is unfortunately quite intrusive because
it required moving a chunk of state from being local to the function to
attributes of the object.

I also nuked the duration-based break condition, since it was never
realistically going to be hit (2 million second limit) and it was
getting more painful to maintain.