Make noise diode check optimistic
If a noise diode fired *at all* within the given time range, then solve
for KCROSS_DIODE. Previously the check only accepted cases where the
noise diode fired continuously within the time range, which is hard to
guarantee with the structure of the calibrate_delays observation script
and therefore caused many "delays not found" errors. It should still
be OK to add in some unwanted noisy dumps, given the signal-to-noise
ratio of the noise diode HV signal. Also do a bit of refactoring and
renaming of the function.