Correctly iterate through antennas when plotting cal_product_G
The report generates plots for 16 antennas at a time, iterating through the total number of antennas in the observation. However for some plots the code failed to allocate a unique name to each iteration of the plot, resulting in the plot being overwritten with each iteration. This has been addressed by appending the index number of the first antenna in the plot to the plot name.

This addresses jira-ticket SR 1159