Use weights in `k_fit`
In a recent pull request NaN's resulting from averaging completely flagged data (with weight zero) have been replaced by zeros in the data.
This creates problems in the delay fitter which previously used `np.isnan` to detect and manage this flagged data is two places:
* in the weights that it inputs into stefcal to predict the bandpass.
* to mask this data when unwrapping the phase and fitting a slope to the phase.

This pull request alters `k_fit` to:
*use the data weights in stefcal, which should be zero for any flagged/zeroed data.
*mask data with `weight=0` when fitting the slope of the phase. This is ensure correct unwrapping of phase even when sections of the data are zero.
*use a weighted least squares fit in the slope fit.

Stefcal does not currently output errors on the bandpass fit to use as weights in the slope fit. As an interim measure the weight for the bandpass is set to zero for any bandpass outputs of NaN and 1 for everything else. This should be updated when stefcal is updated to output errors on the bandpass.