Don't plot flagged data
Currently the cal pipeline sets flagged data to zero. This can be problematic when interpreting the cal report plots of corrected phase as the desired value of the calibrated phase is close to the value of the flagged data.

This commit sets flagged data to NaN prior to plotting, so that it will not appear in plot.