Fix labelling of the referance antenna in tables and plots
Previously every time `make_cal_report` wrote a report it would append a refant label to the reference antenna in the antenna_names list. As antenna_names was only a shallow copy of the parameters[antenna_names] list, this resulted in an extra 'refant' being appended to list for every report written. This has poor readability and also breaks all the table alignments and formatting in the report when this label becomes very long.

This has been fixed by making the 'antenna_names' list a deep copy of the 'parameters[antenna_names]' list.