Fix bug in the calculation of the flags summed over all times
The previous implementation assumed that the pipeline would output a list containing only one array of flags. However the pipeline can output a list of multiple arrays of flags, one per scan processed by the pipeline. Now all scans processed by the pipeline are included in the sum of the flags over time.