Baseline dependent static mask
Only apply the static channel flags on baselines > 1000m. As RFI tends to be stronger
on shorter baselines this strategy masks the worst RFI on the core antennas while allowing
the flagger to flag RFI on longer baselines. This should allow the pipeline to recover an
estimate of the bandpass  on the channels within the static mask provided there is
sufficient good data on long baselines.