Extend flags in frequency
Extending the flags over a larger range in channels seems to produce
a better bandpass solution in 4K mode. There are fewer discrepant points
at the edges of flagged regions. It doesn't appear to be necessary to alter
this for 1K mode or 32K mode but this is pending further investigation.