Prevent unit test from hanging if it fails
If a control test didn't get to the graceful shutdown path for the
servers at the end of the test, then the cleanup code that tries to
shutdown the servers deadlocks. That's because it tries to shut down the
servers serially, but if there is still a capture in progress then there
is an interlock in the shutdown (to agree on the number of dumps in the
last accumulation) and so the first server can't finish shutdown until
the second has started shutdown.

Some extra machinery is added to ensure that the servers are shut down
in parallel.