Fix non-deterministic heap ordering in test
The control tests use multiple endpoints to ensure that katsdpcal can
handle this correctly, but the arrival order of packets between
different endpoints depended on timing (even with the lossless
in-process transport). The packets directed at each server are now all
actually sent to the same endpoint, so that they arrive in a consistent
order.