Add 'Run' parameter to metadata file
This is to address Jira ticket https://skaafrica.atlassian.net/browse/SR-1782, this
parameter is required in order to distinguish the 4 different reports created by split cal
in 32K mode.

Additionally the report names are now appended with the server index, this was requested by users to
make it simpler to immediately know which report you have open.

A unit test for the metadata file has been added.