cal: refactor accumulator towards fixing SR-1812
Most of the Accumulator class is moved into a _CaptureBlock inner class,
which is responsible for handling a single capture block. A capture_done
goes through a few phases:

- It's given a chance to stop on its own (due to stop heaps).
- The receiver is stopped, any block on space in the buffer is
interrupted, and any remaining heaps are discarded. After this, the
_drained_rx_event is set, and capture_done returns to the user. A new
capture-init can immediately be issued.
- The total dump count is synchronised between the cal processes and
extra slots are allocated to align them. This could block for an
extended time waiting for space in the pipeline. After this, the
_done_event is set, which will trigger the subsequent capture block (if
any) to actually start work.