Interpolate across NaNs in B solutions
Applying B solutions with NaNs to the target propagates these NaN's
to the target data which can result in unusual/unexpected flagging
patterns in the target data.

This commit interpolates across NaNs in the frequency axis prior to
applying the solution and thus relies on the target flagger to flag
RFI in the target.