Interpolate NaNs in bandpass across all cal nodes
NaNs in the flagged bandpass propagate to the target data when
the bandpass is applied creating unusual/unexpected flagging
patterns.

This commit therefore stores an interpolated bandpass in
the `solution_store`. All subsequent calls to fetch (and apply)
the bandpass in the pipeline use this interpolated bandpass.

The original flagged bandpass (with NaNs) is still stored in telstate.
This is to preserve information about which regions of the bandpass are suspect.

In the split-cal case all parts of the bandpass are retrieved from telstate
before interpolating. This allows for correct interpolation across the edges of
the cal nodes if there are NaNs at these boundaries.