Use flux model provided in target string
This commit addresses jira ticket https://skaafrica.atlassian.net/browse/SR-893.
Previously the pipeline used only its own internal calibrator models, stored in
/conf/sky_models. This PR allows the pipeline to search for an internal calibrator model,
if one isn't found it will use the model provided by the target string.

To enable this the format of the internal models is modified to better match
the format of  `katpoint.FluxDensityModel`, most notably it now stores model coefficients for the
MHz scale (previously it used the GHz scale). In the future the plan is that katpoint
will be flexible enough to handle both MHz and GHz scales allowing us to store/supply
the co-efficients in the form provided by the original source/reference.

When searching for a model file for a given target the code now searches for a file which matches
the name or the aliases of the target, this helps to slightly robustify the process used to find a model
file.