Get gain solutions on bpcals
Add solving for gains on the bandpass calibrators. These are assumed
to have a flux model (usually in conf) and this is used when deriving
the gains. The gains on the bandpass calibrators are stored in
a new CalSolutionStore labelled 'G_FLUX' in solution_stores. This
is to distinguish them from plan 'G' gains that didn't have a flux model
during their derivation. To allow 'G' gains without models I have added
an option in 'g_sol' in Scan to either apply model or not- then 'g_sol'
on the gain calibrators is run with use_model=False.

Now that we have gains on the bpcals and the gaincals, and a target aware
solution store, I've also added target aware apply calibration. This is
only done for 'G' on the gain calibrators and the bp calibrators, and is
implemented via a 'G_target' kwarg to get_solutions_to_apply. This
defaults to None, which means don't choose which 'G' target to interpolate.