Remove data with invalid weights when averaging
This is to address jira ticket https://skaafrica.atlassian.net/browse/SR-1946, which refers to a
complex divide by zero error in Stefcal. It appears that the cause of these errors are a small number of
data points which have very high weights (> 1e15). These high weights result in an overflow error when
computing a gain during an interation in Stefcal, which ultimately causes it to crash.

To alleviate this the averaging functions within the pipeline now consider data with unfeasibly high
weights (>1e15) to be invalid. These points are discarded during averaging and considered flagged. This should
ensure that averaged data fed into Stefcal will never have these unfeasibly high weights.

Additionally a number of changes have been made to suppress 'divide by zero' warnings. These warnings cause
confusion among users and are often misidentified as the source of unrelated problems in the pipeline. In fact they are usually
benign and are the result of expected behaviour in the pipeline.