Prefer lower case in identifiers
Only class names should start with a capital letter. But more important,
avoid mixing lower and upper case in telstate keys as far as possible,
as that will be stuck in our data sets *forever* (on pain of a katdal
workaround). The flux calibration output is also not a cal product in
the conventional sense, so don't mention `product_F`.