Plot targets with `polcal` tag in report
This addresses jira https://skaafrica.atlassian.net/browse/SR-2014.
The report now includes a plot of the corrected cross-pol phases of all targets tagged with a
`polcal` tag. This produces one plot per scan of a `polcal` calibrator, hopefully this won't be too
onerous as typically there are only a few scans per observation with a `polcal` tagged target.

These plots won't be very useful in the case where the `polcal` target is at a parallactic angle which
degrades the cross-hand signal, in the future it might be useful to include information about the
status of this parallactic angle in the report.