Use frequency dependent parameters in cal
This addresses https://skaafrica.atlassian.net/browse/SR-1692 and allows the pipeline to have
separate parameters for UHF, L and S band.
Previously frequency based cal pipeline parameters were supplied in channels either via a config file or as
an optional argument. Now they are supplied in either MHz or Hz and converted to channels before
being stored in telstate and sent to the pipeline process.

The rfi mask was previously supplied as a pickle file with the correct number of channels, it is now supplied
as a text file specifying the range of frequencies to mask, which is then broadcast to channel frequencies of
the observation.