Add test for resetting the solution stores
The test sends two capture blocks with two different target names
to the pipeline. It checks that the solution stores contain
the target associated with each capture block and not the other
after the pipeline is finished for each capture clock. It also checks
that the correct targets are contained in the 'measured_flux'
telstate key for each capture block.