Include Fluxcal info in report
The report now provides a table of measured fluxes of gain calibrators as provided in
`telstate['measured flux']`. A model value is included in the table for comparison if a model is provided.

Additionally if a flux model exists for a gain calibrator it is now plotted in all plots of amplitude versus frequency.

In all plots of the gain calibrators the visibilities are scaled to the measured fluxes reported in telstate[`measured_flux`].

The target data is not gain calibrated, as the flux of the target is unlikely to be known `a priori` this doesn't really have
a negative impact on the diagnostic value of the target plots in the report, and avoids the need to introduce a complicated scheme
to correctly interpolate gains between different gain calibrators onto the target