Use frequency-valued parameters
This addresses JIRA ticket SR-1692 and allows the pipeline to have
separate parameters for UHF, L and S band. Previously frequency-based
cal pipeline parameters were supplied in channels in the config files.
Now they are supplied in either MHz or Hz and will be converted to
channels before being stored in telstate.

The S-band and UHF-band files start with the same channel range as
L-band (i.e. channel 2250-2450 for a 4k channelisation), expressed in
Hz for the specific receiver band.