Handle bands without a HV phase 'to_sky' spline
For bands which have no measured 'to_sky' phase spline correction
a placeholder spline with co-efficients of zero is added to the config file.

The pipeline checks that there are non-zero co-efficients in the 'to_sky' spline
before producing a 'BCROSS_DIODE_SKY' product. No product is produced for the case of
all zero co-efficients