Specify flux models with wsclean polynomials
This is to address https://skaafrica.atlassian.net/browse/MKAIV-1531.
In order to support more complex models of calibrator fields it is
necessary to expand the range of polynomials used to specify flux density
models in cal.

This PR specifies a custom FluxDensityModel class which supports
specifying a model as either a Baars et al. 1977 polynomial or either of
the two formats used by wsclean