Address review comments
The format of the description string for `wsclean` model components has
been reworked to be closer to the output of `wsclean`. The accepted
format is now of the form:
`min_freq, max_freq, wsclean I [coef_1 coef_2 ... coef_5] false/true
ref_freq`

Minimum and maximum frequencies are supplied in units of MHz, while the
reference frequency is supplied in units of Hz.