Improve the models plotted in the cal report
Previously the report only reported and plotted the flux from the first
component (assumed to be dominant in the field) in the calibrator field
model.

This PR updates this to plot the average (across baseline and time) of
all the components in the calibrator field model.

To achieve this the pipeline saves the averaged model in the corrected
data dictionary which is passed to the report writer at the end of the
observation. This allows the model to be calculated only once, rather
than having to recalculate it from scratch in the report writer.

In the split cal case, the report writer for the server
containing the relevant channels for the flux calculation (ie the g_gchan
to g_echan interval) calculates the average flux for the interval and
stores it in telstate with the key 'target_name_model_flux'.