Increase timeouts for failing tests
Periodically some of the tests of `control.py` are taking longer to
run than the timeout on the shutdown requests. This causes the tests to
fail because the report writer tries to finish writing the report to a
directory that has been deleted during cleanup.

I'm increasing the timeouts to reduce the likelihood of these tests
failing.