Support multi-frequency setups for S-band
This is to address https://skaafrica.atlassian.net/browse/SPR1-463.
The cal pipeline uses a narrow range of frequencies (specified by e.g.
g_bfreq, g_efreq) to perform gain and delay calibration. These have
always been provided as a single float per band. S-band has multiple (5)
possible frequency setups across the full 1750 - 3500 MHz band, and a
single frequency range cannot be used for all 5 possible bands.

This modifies the pipeline to allow these parameters to be supplied as
either a list of a float. If a list is provided another parameter
('band_start_freq') is required, which provides a list of starting
 frequencies that correspond to the list of gain and delay calibration windows.