Major rework of store-inplace
The old implementation would materialise the graphs of the sources as a
dictionary, which probably causes all sorts of inefficiencies, and might
be why we're encountering https://github.com/dask/dask/issues/7632. I've
rewritten it to just extract the numpy arrays from the targets, and use
a wrapper class together with `da.store` so that the actual computation
uses the built-in Dask machinery instead of graph surgery.