Yet another rework of store_inplace
This time it actually works with dask.distributed (and this is tested).
I don't really know why the previous version didn't - but since this is
all void-the-warranty use of dask I'm not too worried. This returns
mostly to the original implementation (building a new layer by hand,
instead of using dask.store), but it works with HighLevelGraph instead
of squashing all the layers.