Select calibration windows for narrowband
As described in JIRA SPR1-1058 narrowband observations don't have
a fixed centre frequency. It is thus not possible to select a fixed
frequency interval for gain and delay solutions as is done for
wideband.

To accomodate this a new parameter file is created for narrowband
which leaves the solution interval parameters (e.g. `g_bfreq`) blank.
The pipeline will automatically select and set these solution interval
parameters if they are blank.

It achieves this by selecting the largest consecutive group of channels
that will fit within one server and is not masked by the static RFI and band
edge masks. It selects a fixed number of channels from within this interval
as the solution interval.

The channel interval is chosen to be half the width of the bandwidth
used in wideband mode. This is because the full width would span more
than one calibration server and thus cannot be accomodated using the
current multi-server implementation.