Generate report even when no refant is selected
This is to address SPR1-1117, certain observations (mainly TRAPUM
targets) do not observe any calibrators, only targets. The pipeline is
only triggered to select a reference antenna when a calibrator target is
observerd so for these observations no reference antenna is selected.

This PR fixes a problem with internal logic of generating a report when
no reference antenna is selected.