Fix pipeline failures caused by `polcal` tag
This is to address SPR1-1081. If a target was tagged as `polcal` the
pipeline accidently overwrote the `data` variable containing the current buffered
data to be processed with the scan summary data that is written to the
calibration report. Subsequent attempts by the pipeline to access any
unprocessed scans in the `data` buffer then result in an exception.

This has been remedied by giving the scan summary tuple a unique
name (i.e. not `data`).