Handle case when calibrators are tagged as targets
SPR1-1301, some users add a `target` tag to their calibrators, this
caused the calibration report to only generate `target` style plots
in the report. This PR modifies the report to only generate the
`calibrator` style plots in these cases, which usually contain both
phase and amplitude information (instead of just amplitude for
`targets`) for better insight into the calibration success on these
calibrators.