Update weights when pre-applying solutions
SPR1-1316. Pre-applying calibration solutions to visibilities without
updating the visibility weights results in antennas with
poor visibilities/solutions being upweighted in the following calibration
solution. This can result in the solver producing poorer solutions for good antennas at the
expense of trying to fit the poor antenna. This PR updates the weights
appropriately if solutions are applied prior to solving for a
calibration product.