Keep model visibility dimensions consistent
If no model is available for a given source the model was previously set
to np.array([1]), but this has a different dimension to the model
visibilities produced when a model exists. Processing of
model visibilities to include in the cal report assumes that they have a dimensions (ntime,
nchans, npols, nbaselines), thus when no model is present an exception
was produced. This is remedied by ensuring the model visibilities have a
consistent set of dimensions.