SPR1-1869.Pre_g solutions are relative to the mean
Previously pre_g gain solutions were altered to be relative to the first
solution per antenna. This was intended to make the solutions more
robust to poor antennas, while still allowing the pre_g gain solution to
remove any gain drift across the calibrator scan.

Unfortunately this makes the solutions very sensitive to the quality of
the first dump in the scan. If this dump is flagged (as can easily
happen if an antenna is straggling) then all solutions relative to this
flagged (NaN'ed) solution will become NaN. To prevent this the solutions
are now relative to the mean solution for scan interval.