Interpolate across NaN gain solutions
As seen in SOQ-3 if a gain solution is a NaN for any reason, when this
solution is interpolated and applied to a target scan it will cause that
scan to be completely flagged by the cal pipeline. This PR allows the
pipeline to ignore these NaN gain solutions and interpolate from the
nearest valid gain solution.