Fix failing tests
Relax the constraints in test_capture_complex_models and the new refant selection selects a longer baseline antenna introducing more noise into the solutions.

Modify the refant selection to be >= median-mad, (instead of >), this allows a valid refant selection even when the data is identical on all baselines, which is the case from some of the tests (e.g. test_reset_solution_stores).

Check if the old antenna is selected by the new refant selection method before removing it from the refant candidate list. This allows set_refant to successfully find a new reference antenna if the old one is flagged.