SPR1-3111 Replace nose asserts with plain asserts
Run the nose2pytest script on the unit test directory, which rewrites
most of the calls to `assert_*` functions imported from `nose.tools`
and replaces them with equivalent vanilla assert statements.

The script doesn't (yet) run on Python 3.12 so I ran it on Python 3.11.

Afterwards I swapped the order of the operands of some of the equality
expressions, since I prefer "assert measurement == expected value".
For this I used a regex search and replace in VS Code to assist me:
  - search: assert (.+) == (.+)
  - replace: assert $2 == $1