SPR1-3111 Replace asynctest with unittest 3.8
This uses the IsolatedAsyncioTestCase class introduced in Python 3.8
instead of asynctest, which won't run on Python >= 3.11 anymore.

The alternative is pytest-asyncio but that would involve untangling the
classes and turning them into fixtures instead - a somewhat non-trivial
endeavour.

Note that the package still potentially runs on Python 3.5 but the tests
now require Python 3.8. This is hardly an issue since the base docker
image uses Python 3.8 anyway, and everyone else has moved on.

The dask issues described in SPR1-1790 still prevent katsdpcal from
running on Python >= 3.11 though... But at least we now have the ability
to run the tests to see why they fail.