SPR1-2993 Streamline telstates, reduce hard-coding
We've made an L0 stream view for telstate and stored it as
`self.telstate_l0` on the `TestCalDeviceServer` object, so why not
use it? This avoids explicit use of the "sdp_l0test" stream name and
is also more compact.

Furthermore, store the dump period and first timestamp as attributes
to avoid having to look it up in telstate, or worse, hard-coding it
explicitly. The first timestamp is especially tricky to look up since
it resides in a composite "cb_sdp_l0test" namespace, which is why there
are many explicit "100.0"s floating around. This should make the tests
more understandable too by describing how values are constructed.