SPR1-3025 Add explicit parts count for EPOINT
The EPOINT product could have used `product_B_parts` as well, but that
would require a custom lookup somewhere. Katdal looks up the parts based
on a generic product name and benefits from having these explicit. It
knows to look for parts if the corresponding `product_*_parts` key is
found.