SPR1-3025 Set EPOINT soltype and ignore soltarget
The EPOINT product has a custom solution type (5 parameters), not B.
Ignore target name stored in B_POINTING solutions because of the way
CalSolutionStore is implemented. It is not possible to preserve the
target without first knowing what it is, so the `b_solutions.target`
is always None. It only affects the logger output and will be replaced
by a proper target in the EPOINT solutions. There is little benefit
to check the target anyway, since we wipe the B_POINTING store after
each scan and we expect multiple scans on the same target anyway.