SPR1-1790 Allow cal to work with newer dask
The latest version of dask + distributed that allows katsdpcal to function properly is 2021.9.0. In more recent versions of dask the `store_inplace` functionality is broken.

`store_inplace` allowed the pipeline to write changes to the numpy arrays backing the dask arrays used in the pipeline processing. In order do away with this functionality this PR introduces the following changes

1.) Instead of `apply_inplace` or `store_inplace` the pipeline computes the outcome of the required computation and recreates a new dask array based on these computed outcomes

2.) The `store_inplace` functionality allowed the pipeline process to store its computed flags in the same flag buffer array that was also accessed by the flag writer process. Instituting the changes in point 1.) above meant that the flag buffer was no longer being updated to include the computed cal `l1` pipeline flags. The flag buffer array is now explicitly passed to the pipeline function in reduction.py and manually updated with the computed l1 flags.