correct sync_time in simulator
The slew scanstates were been sent into the pipeline for correction  they were been considered as
tracks by the previous code. This was due to adding the 'correct' sync_time to the simulator and
this resulted in the timestamps being incorrected handel. This commit adds a fix to the simulator.py
by subtracting the sync_time from the timestamps to keep the times aligned with an observation.