Fix for test_flush_pipeline_exception
In this commit we add a fix for the failing test_control.py test.
The test inject exceptions into the pipeline. For our case the test
failed because the test telstate_cb does not have the stream type
`sdp.vis` for the `check_applied_gain()` to extract sensor information
from Katdal. We bypass the test by adding a target with a `pointingcal`
tag and added this target to the telstate, this will ensure that the
test checks for the reference pointing solutions are flushed from the
pipeline. Phase up observations have a `bfcal` tag. The test will
therefore check exceptions for observations that have a reference pointing tag.

In this commit we also debug the plotting of the tracks, to ensure that
in each track the respective nmad phase is plotted. The issue here was
the name was duplicated for the plots and this resulted in the same plot
appearing in the cal_report for the different tracks.