Update `simulator.py`
In this ticket we have address the use of the  katdal file to generate
pipeline data for the simulator. In this commit we remove the `SimDataMS` subclass,
we remove the factory method that use to construct the data object based on the subclass
Now we rely only on katdal file, so its best to remove the factory()
method. Further in this commit we move out the `get_params` from the
class to promote better readibility, the same goes for the `write_data`
function (write data to file) this is currently not been called here.

We have also cleaned up our populating of telstate with
immutabels/mutables to make it compact.