SPR1-3225-fix-failing-cal-report-generation
The report generation failed for non-phase up observations. In this
commit we add a try-except clause to check wether the av_corr('nmad_phase') data contains valid data for plotting, this is added to the `write_phase_stability`
function in `report.py`. Alternatively, we can try to make a more strict
conditional statements in the `reduction.py`, currently in the code
checks if its phase up and checks for the corrected track and calculated
the nmad_phases and stores this in the sumarise stats, for the non-phase
up case it rightly does not calulate the nmad_phases but the nmad_phase
tag is still created and it this means the av_corr('nmad_phase') data is
empty.

Having the try-except in the report is good, because this will check
data validity and will continue/ skip to the next plot generations if
invalid data is present (as in the case of non-phase ups)