SPR1-3249-fix_check_applied_gain
In this commit we address the failing nmad phase plot generation due to
sensor keys that are not available in the live telstate of an
observation. We used found aliases of `stream_name', `obs_label` and
`capture_block_id` to be used by `check_applied_gain` to make the
required katdal object for us to get the applied gain values to check if
the observation is phase up to trigger the pipeline to make nmad phase plots.

We also address the issue of the pipeline triggering nmad phase
generation for image observations, this should not happen. Here we move
the `bfcal` check as the first check, to flag and pass only observations
that have the `bfcal` tag.