Remove deprecated functionality
The preferred syntax for `async_timeout.timeout` was modified to `async with
async_timeout.timeout` in version 1.4.0 and the older syntax is no longer supported in
newer versions.

`np.product` has been deprecated since version Numpy 1.25.0 and should be
replaced with `np.prod` in future.

Numpy version 2.0.0 introduced a breaking change that altered how the
precision of scalars are preserved when combining dissimilar data types.
Some of the tests were failing due to this change and required explicity casting to
a np.uint8 type to pass the test.

Since Python 3.10 using `asyncio.get_event_loop()` when there is no
running loop raises a runtime warning. To avoid this explicitly create
and set the event loop.

Newer versions of dask raise a large number of spurious warnings about
large dask graph sizes so a config parameter is set here to suppress
these.